+++ inst/dummypackage/tests/testthat/test-addition.R
@@ -1,0 +1,3 @@
+test_that("Add works", {
+  expect_equal(1 + 1, 2)
+})