+++ NEWS.md
@@ -1,12 +1,12 @@
-# attachment (development version)
+# attachment 0.3.0
 
 ## New features
 
-- `set_remotes_to_desc()` now detects github, gitlab, git, local installations (@MurielleDelmotte)
+- `find_remotes()` and `set_remotes_to_desc()` now detects github, gitlab, git, bitbucket, local installations to add to the "Remotes:" part of the DESCRIPTION file (@MurielleDelmotte)
+- Quarto documents can be parsed with `att_from_qmds()` as an alias of `att_from_rmds()`.
 
 ## Minor changes
 
-- Quarto documents can be parsed with `att_from_rmds()`.
 - Documentation for bookdown and quarto dependencies extraction updated
 
 ## Bug fixes